backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
passlib[bcrypt]
sqlalchemy
